% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoDxc.R
\name{calc.mz.Window}
\alias{calc.mz.Window}
\title{Function to calculate m/z window given an ion and ppm tolerance}
\usage{
calc.mz.Window(mz, ppm)
}
\arguments{
\item{mz}{m/z}

\item{ppm}{ppm tolerance for the m/z window}
}
\value{
A vector of length 2 with min m/z and max m/z
}
\description{
Function to calculate m/z window given an ion and ppm tolerance
}
