% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot2D}
\alias{plot2D}
\title{Plot a 2D Chromatogram}
\usage{
plot2D(
  object,
  file = 1,
  mod.time,
  delay.time = 0,
  ion = NULL,
  mz.tol = c("ppm", "abs"),
  ppm.tol = 20,
  abs.tol = 0.5,
  log.scale = F,
  rt.min = 0,
  rt.max = max(rtime(object)),
  rt2.min = 0,
  rt2.max = mod.time,
  mz.min = min(object@featureData$lowMZ),
  mz.max = max(object@featureData$highMZ),
  int.min = 0,
  int.max = max(tic(object), na.rm = T) + 1,
  color.scale = "Spectral",
  reverse.scale = T,
  save.output = F,
  filename = "2dplot",
  filetype = ".pdf",
  filepath = ".",
  print.output = T,
  mz.digits = 4,
  plot.type = c("2D", "2Di", "3D")
)
}
\arguments{
\item{object}{An MSnExp object}

\item{file}{Which file (numeric, not filename) to pull data from. Default is
1}

\item{mod.time}{The modulation time of the 2D run}

\item{delay.time}{The time it takes for modulation to start.
Default is 0.}

\item{ion}{Select an ion to generate an extracted ion chromatogram (EIC). If
missing, the total ion chromatogram (TIC) is returned.}

\item{mz.tol}{Whether EIC m/z range is calculated using ppm (relative) or
absolute tolerance. Default is ppm}

\item{ppm.tol}{PPM tolerance, used to calculate m/z range for EIC. Default is
20}

\item{abs.tol}{Absolute tolerance, used to calculate m/z range for EIC.
Default is 0.5 m/z}

\item{log.scale}{Whether intensity coloring should be log scaled. Default is
FALSE}

\item{rt.min}{Minimum 1D RT to plot. Default is 0}

\item{rt.max}{Maximum 1D RT to plot. Default is the max found in the data}

\item{rt2.min}{Minimum 2D RT to plot. Default is 0}

\item{rt2.max}{Maximum 2D RT to plot. Default is the modulation time}

\item{mz.min}{Minimum m/z to plot. Default is lowest found in the data}

\item{mz.max}{Maximum m/z to plot. Default is highest found in the data}

\item{int.min}{Minimum intensity to plot. Default is 0}

\item{int.max}{Maximum intensity to plot. Default is highest found in the TIC}

\item{color.scale}{Color scale from distiller to use. Default is 'Spectral.'}

\item{reverse.scale}{Reverse the color scale. Default is T.}

\item{save.output}{Whether to save the output as a file. Default is FALSE}

\item{filename}{The name for saving the file if desired. Default is
'2dplot.png' plus the file selected plus the ion selected.}

\item{filetype}{Filetype of the form ".pdf". Default is ".pdf"}

\item{filepath}{Where to save the output file. Default is '.'}

\item{print.output}{Whether to show the plot output. Default is TRUE}

\item{mz.digits}{The number of digits after the decimal to show for m/z's
in the plot title. Default is 4}

\item{plot.type}{Choose from 2D ggplot, interactive 2D plotly plot ('2Di'),
or an interactive 3D plotly plot ('3D'). Default is 2D}
}
\value{
A 2-dimensional plot
}
\description{
This function takes an MSnExp object and plots a 2D chromatogram using the
specified modulation time and optional ion extraction. The function relies on
ggplot2 for plotting.
}
